requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==21.6b0
click==8.0.1
Django==3.0
django-common-helpers==0.9.2
django-cors-headers==3.7.0
django-countries==7.0
django-filter==2.4.0
django-js-asset==1.2.2
django-phonenumber-field==5.0.0
djangorestframework==3.10.0
djangorestframework-simplejwt==4.4.0
mypy-extensions==0.4.3
pathspec==0.8.1
phonenumbers==8.12.16
Pillow==8.2.0
PyJWT==1.7.1
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
